add fileSize to modVersions
This will allow the frontend and installers to show the size of a mod. Can also allow them to show a warning for oversized mods.

Warning: this commit requires database migration code in order not to completely break things!

Note: in updateMod, fileSize is required to work, and not error out. However, it is only set by the modVersion's fileSize, not ever the user.
Then in approval, where fileSize is also required, it is pulled from the modVersion's fileSize in the queue. Of which was earlier set from the modVersion's fileSize. fileSize should not ever be able to be set by a user, only by the program, originating from the uploaded version's file size